+++ e2e-tests/codegen-tests/tests/parser.test.ts
@@ -20,4 +20,7 @@ it("works", () => {
   expect(Codecs.AliasToConst.describe()).toMatchInlineSnapshot(
     '"type CodecAliasToConst = \\"constant value\\";"',
   );
+  expect(Codecs.TestHoist.describe()).toMatchInlineSnapshot(
+    '"type CodecTestHoist = { a: Array<string>, b: Array<string> };"',
+  );
 });