+++ packages/beff-wasm/fixtures/codegen-snaps/export-decoder/bff-generated/parser.js
@@ -1004,6 +1004,27 @@ class ParserRefImpl {
     return to.reportDecodeError(ctx, input);
   }
 }
+class ParserHoistedImpl {
+  hoistedIndex;
+  constructor(hoistedIndex) {
+    this.hoistedIndex = hoistedIndex;
+  }
+  describe(ctx) {
+    return hoistedIndirect[this.hoistedIndex].describe(ctx);
+  }
+  schema(ctx) {
+    return hoistedIndirect[this.hoistedIndex].schema(ctx);
+  }
+  validate(ctx, input) {
+    return hoistedIndirect[this.hoistedIndex].validate(ctx, input);
+  }
+  parseAfterValidation(ctx, input) {
+    return hoistedIndirect[this.hoistedIndex].parseAfterValidation(ctx, input);
+  }
+  reportDecodeError(ctx, input) {
+    return hoistedIndirect[this.hoistedIndex].reportDecodeError(ctx, input);
+  }
+}
 const buildParsers = (args) => {
   const stringFormats = args?.stringFormats ?? {};
   for (const k of RequiredStringFormats) {
@@ -1108,13 +1129,15 @@ const buildParsers = (args) => {
 
 const RequiredStringFormats = ["password","StartsWithA"];
 const RequiredNumberFormats = [];
+const direct_hoist_0 = new ParserTypeOfImpl("string");
+const hoistedIndirect = [];
 const namedParsers = {
     "User": new ParserObjectImpl({
         "age": new ParserTypeOfImpl("number"),
-        "name": new ParserTypeOfImpl("string")
+        "name": direct_hoist_0
     }, null),
     "NotPublic": new ParserObjectImpl({
-        "a": new ParserTypeOfImpl("string")
+        "a": direct_hoist_0
     }, null),
     "StartsWithA": new ParserStringWithFormatImpl([
         "StartsWithA"