+++ packages/beff-client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @beff/client
 
+## 0.0.102
+
+### Patch Changes
+
+- smart hoist codegen
+- Updated dependencies
+  - @beff/cli@0.0.102
+
 ## 0.0.101
 
 ### Patch Changes